Time: 7 ms (83.62%) | Memory: 9 MB (82.09%) - LeetSync

+++ 3-longest-substring-without-repeating-characters/longest-substring-without-repeating-characters.cpp
@@ -6,7 +6,7 @@ public:
         string holder;
 
         for (int i=0;i<s.size();i++){
-            size_t pos = holder.find(s[i]);
+            int pos = holder.find(s[i]);
             if (pos == -1) {
                 holder += s[i];
                 longest = max(longest, static_cast<int>(holder.size()));
